Converted the .NET Framework demo app to basic MVVM
